Pulled the populate to utils and added a test
